refactor: move CheckStatus to common/verifier
