+++ src/common/hooks/useDocumentVerifier.test.tsx
@@ -2,7 +2,7 @@ import sampleDoc from "../../../fixtures/demo-oc.json";
 import { useDocumentVerifier } from "./useDocumentVerifier";
 import { SignedDocument } from "@govtechsg/open-attestation";
 import { renderHook } from "@testing-library/react-hooks";
-import { CheckStatus } from "../../constants/verifier";
+import { CheckStatus } from "../verifier";
 
 import { checkValidity } from "../../services/DocumentVerifier";
 jest.mock("../../services/DocumentVerifier");
