+++ src/components/DocumentRenderer/ValidityBanner/ValidityBanner.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, wait } from "@testing-library/react-native";
 import { ValidityBanner } from "./index";
-import { CheckStatus } from "../../../constants/verifier";
+import { CheckStatus } from "../../../common/verifier";
 
 jest.useFakeTimers();
