--- src/constants/verifier.ts
@@ -1,5 +1,0 @@
-export enum CheckStatus {
-  CHECKING = "CHECKING",
-  VALID = "VALID",
-  INVALID = "INVALID"
-}
